% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasFieldCorrect.R
\name{biasFieldCorrect}
\alias{biasFieldCorrect}
\title{A Three-stage Procedure for Multimodal Brain Tumor Segmentation}
\usage{
biasFieldCorrect(infile, outdir, iter = 100, res = 3)
}
\arguments{
\item{infile}{Full path of original MR image.}

\item{outdir}{The directory for output file.}

\item{iter}{Iterations to run.}

\item{res}{Level of resolution.}
}
\description{
\code{mbts} analyzes T1ce, FLAIR and T2 images and segments the images 
into brain tissues.
}
\author{
Hongda Zhang
}
