% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShift.R
\name{getShift}
\alias{getShift}
\title{Get Shift for Unary Potentials}
\usage{
getShift(res, m)
}
\arguments{
\item{res}{Matrix of estimated mean, variance and interaction coefficients.}

\item{m}{Threshold for bright signals.}
}
\description{
\code{getShift} get the shift for unary potentials.
}
\author{
Hongda Zhang
}
