% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbts.R
\name{mbts}
\alias{mbts}
\title{A three-stage procedure for brain tumor segmentation}
\usage{
mbts(
  patient,
  out = "SEG",
  infolder = "N4ITK433Z",
  shrink = FALSE,
  delta = list(t1ce = c(-1, 0, 8, 4), flair = c(-0.45, 0, NA_real_, 4), t2 = c(2.65, 0,
    NA_real_, 4), fthr = c(0, 0, 4, 5)),
  delta_factor = list(t1ce = 1.75, flair = 2.6, t2 = 4.7),
  gamma = list(t1ce = 0.8, flair = 0.4, t2 = 0.4, fthr = 0.8),
  alpha = list(t1ce = c(10, 10, 10, 10), flair = c(10, 10, 10, 10), t2 = c(10, 10, 10),
    fthr = c(10, 10)),
  beta = list(t1ce = c(1, 1, 1, 1), flair = c(1, 1, 1, 1), t2 = c(1, 1, 1), fthr = c(1,
    1)),
  lambda2 = list(t1ce = c(1, 1, 1, 1), flair = c(1, 1, 1, 1), t2 = c(1, 1, 1), fthr =
    c(1, 1)),
  a = 2,
  nu2 = list(t1ce = rep(0.25, 3), flair = rep(0.25, 3), t2 = c(0.25, 0.25), fthr =
    rep(0.25, 2)),
  maxit = list(t1ce = 10L, flair = 1L, t2 = 1L, fthr = 40L),
  min_enh = 500L,
  min_enh_enc = 2000L,
  max_prop_enh_enc = 0.1,
  max_prop_enh_slice = 0.2,
  min_tumor = 2000L,
  spread_add = 10,
  spread_rm = 9,
  trim1_spread = 10,
  trim1_round = 18,
  remove2d_spread = 25,
  remove2d_round = 25,
  spread_trim = 14,
  round_trim = 20,
  on_flair_prop = 1.5,
  on_flair_hull_prop = 0.3,
  on_flair_nt_prop = 0.3,
  last_rm_solidity = 2,
  last_rm_spread = 16,
  last_rm_round = 16,
  last_trim_spread = NULL,
  last_trim_round = NULL,
  last_trim_rm_spread = 2,
  last_trim_rm_round = 10000,
  csf_check = 0L
)
}
\description{
\code{mbts} Analyzes T1ce, FLAIR and T2 images and segments the images 
into brain tissues.
}
