% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Get Current Threshold}
\usage{
threshold(res, shift)
}
\arguments{
\item{res}{Matrix of estimated mean, variance and interaction coefficients.}

\item{shift}{Current shift for unary potentials.}
}
\description{
\code{threshold} gets current threshold for bright signals.
}
\author{
Hongda Zhang
}
