% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post}
\alias{post}
\title{Postprocessing Stage of the Procedure}
\usage{
post(
  patient,
  out = "SEG",
  infolder = "N4ITK433Z",
  delta = list(t1ce = c(-1, 0, 8, 4), flair = c(-0.45, 0, NA_real_, 4), t2 = c(2.65, 0,
    NA_real_, 4), fthr = c(0, 0, 4, 5)),
  delta_factor = list(t1ce = 1.75, flair = 2.6, t2 = 4.7),
  gamma = list(t1ce = 0.8, flair = 0.4, t2 = 0.4, fthr = 0.8),
  alpha = list(t1ce = c(10, 10, 10, 10), flair = c(10, 10, 10, 10), t2 = c(10, 10, 10),
    fthr = c(10, 10)),
  beta = list(t1ce = c(1, 1, 1, 1), flair = c(1, 1, 1, 1), t2 = c(1, 1, 1), fthr = c(1,
    1)),
  lambda2 = list(t1ce = c(1, 1, 1, 1), flair = c(1, 1, 1, 1), t2 = c(1, 1, 1), fthr =
    c(1, 1)),
  a = 2,
  nu2 = list(t1ce = rep(0.25, 3), flair = rep(0.25, 3), t2 = c(0.25, 0.25), fthr =
    rep(0.25, 2)),
  maxit = list(t1ce = 10L, flair = 1L, t2 = 1L, fthr = 40L),
  min_enh = 500L,
  min_enh_enc = 2000L,
  max_prop_enh_enc = 0.1,
  max_prop_enh_slice = 0.2,
  min_tumor = 20000L,
  spread_add = 8,
  spread_rm = 9,
  trim1_spread = 8,
  trim1_round = 17,
  remove2d_spread = 25,
  remove2d_round = 25,
  spread_trim = 7,
  round_trim = 16,
  on_flair_prop = 1.5,
  on_flair_hull_prop = 0.3,
  on_flair_nt_prop = 0.3,
  last_rm_solidity = 2,
  last_rm_spread = 16,
  last_rm_round = 16,
  last_trim_spread = NULL,
  last_trim_round = NULL,
  last_trim_rm_spread = 2,
  last_trim_rm_round = 10000,
  csf_check = 0L
)
}
\arguments{
\item{patient}{A vector of file names of FLAIR, T1, T1ce and T2 images.}

\item{out}{The folder to save the intermediate and final results.}

\item{infolder}{The folder which includes the multimodal images.
in each dimension of 3D space.}

\item{delta}{A list of values related to unary potentials for 
segmenting T1ce, FLAIR and T2 images and for spliting non-enhancing 
tumor and edema.}

\item{delta_factor}{A factor for adjusting unary potentials. The larger the
values, the smaller the resulting unary potentials.}

\item{gamma}{A list of parameters for pairwise potentials. The larger the 
values the stronger the spatial correlation between neighboring voxels.}

\item{alpha, beta}{Shape and scale parameter of inverse-gamma 
prior of variances.}

\item{lambda2}{Prior variance of interaction parameters.}

\item{a}{Shape parameter of prior distribution of mean intensity of 
bright signals.}

\item{nu2}{Variance of mean intensity of distinguishable classes.}

\item{maxit}{Maximum iterations to run for getting the segmentation results.}

\item{min_enh}{Minimum number of enhancing tumor voxels 
a high-glioma patient has.}

\item{min_enh_enc}{Minimum number of enhancing tumor voxels enclosed in
tumor regions a high-glioma patient has.}

\item{max_prop_enh_enc}{Maximal proportion of enhancing tumor voxels 
enclosed in tumor regions.}

\item{max_prop_enh_slice}{Maximal proportion of enhancing tumor voxels 
enclosed in a 2D slice of tumor regions.}

\item{min_tumor}{Minimal size of a tumor region.}

\item{spread_add}{The maximal form factor a region to be added can have.}

\item{spread_rm}{The minimal form factor a region to be removed can have.}

\item{trim1_spread, trim1_round}{Form factor and roundness for trimming the 
regions for the first time.}

\item{remove2d_spread, remove2d_round}{2D form factor and roundness for 
removing 2D slices.}

\item{spread_trim, round_trim}{Form factor and roundness for trimming the 3D 
regions.}

\item{on_flair_prop, on_flair_hull_prop, on_flair_nt_prop}{Proportions for 
extending whole tumor regions in FLAIR images.}

\item{last_rm_solidity, last_rm_spread, last_rm_round}{Solidity, form factor
and roundness for removing the 3D tumor regions for the last time.}

\item{last_trim_spread, last_trim_round}{Form factor and roundness for
trimming the tumor regions for the last time.}

\item{last_trim_rm_spread, last_trim_rm_round}{Form factor and roundness
for removing tumor regions after trimming the tumor regions for the 
last time.}

\item{csf_check}{Whether to validate some of the necrosis 
voxels are actually CSF.}
}
\description{
\code{post} postprocesses the images.
}
\author{
Hongda Zhang
}
